pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py313"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "D",  # pydocstyle
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.poetry]
name = "reddit-weekly-top"
version = "0.1.0"
description = "A tool to fetch top weekly posts from selected subreddits"
authors = []
packages = [
    { include = "reddit_weekly_top", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = ">=1.0.0"
pydantic = ">=2.5.3"
praw = ">=7.7.1"
supabase = "^2.3.0"
langchain = "^0.3.0"
langchain-anthropic = "^0.2.0"
langchain-community = "^0.3.0"
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.3.0"
requests = ">=2.31.0"
yt-dlp = ">=2024.7.25"
faster-whisper = ">=1.0.3"
google-api-python-client = {version = ">=2.173.0,<3.0.0", optional = true}
langsmith = "^0.4.4"
flask = "^3.1.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.4"
pytest-mock = ">=3.12.0"
pytest-cov = ">=4.1.0"
ruff = ">=0.1.14"
mypy = ">=1.8.0"
streamlit = "^1.46.0"

[tool.poetry.scripts]
reddit-fetch-posts = "reddit_weekly_top.scripts.fetch_posts:main"